// frontend/src/api/models/TrimConfigurationUI.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Klippers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TrimConfiguration } from './TrimConfiguration';
import {
    TrimConfigurationFromJSON,
    TrimConfigurationFromJSONTyped,
    TrimConfigurationToJSON,
    TrimConfigurationToJSONTyped,
} from './TrimConfiguration';

/**
 * 
 * @export
 * @interface TrimConfigurationUI
 */
export interface TrimConfigurationUI {
    /**
     * 
     * @type {TrimConfiguration}
     * @memberof TrimConfigurationUI
     */
    trimConfiguration?: TrimConfiguration;
    /**
     * Video url.
     * @type {string}
     * @memberof TrimConfigurationUI
     */
    videoUrl?: string;
    /**
     * Thumbnail url.
     * @type {string}
     * @memberof TrimConfigurationUI
     */
    thumbnailUrl?: string;
}

/**
 * Check if a given object implements the TrimConfigurationUI interface.
 */
export function instanceOfTrimConfigurationUI(value: object): value is TrimConfigurationUI {
    return true;
}

export function TrimConfigurationUIFromJSON(json: any): TrimConfigurationUI {
    return TrimConfigurationUIFromJSONTyped(json, false);
}

export function TrimConfigurationUIFromJSONTyped(json: any, ignoreDiscriminator: boolean): TrimConfigurationUI {
    if (json == null) {
        return json;
    }
    return {
        
        'trimConfiguration': json['trim_configuration'] == null ? undefined : TrimConfigurationFromJSON(json['trim_configuration']),
        'videoUrl': json['video_url'] == null ? undefined : json['video_url'],
        'thumbnailUrl': json['thumbnail_url'] == null ? undefined : json['thumbnail_url'],
    };
}

export function TrimConfigurationUIToJSON(json: any): TrimConfigurationUI {
    return TrimConfigurationUIToJSONTyped(json, false);
}

export function TrimConfigurationUIToJSONTyped(value?: TrimConfigurationUI | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'trim_configuration': TrimConfigurationToJSON(value['trimConfiguration']),
        'video_url': value['videoUrl'],
        'thumbnail_url': value['thumbnailUrl'],
    };
}